Support ejabberd 18.04
Work with ejabberd 18.04. The main difference is a change from using
3-tuple for the xmpp packet to a record.

This version keeps the behaviour of returning 'drop' if the packet is
denied. Other mods return {stop, drop}. Presumably these hooks are
called via a fold and {stop, drop} will cause early termination.

Lightly tested, seems to be working.